fix(ax1): isolate sqlite migration support
